requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytesseract==0.3.10
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
PyPDF2==3.0.1
pdfplumber==0.10.3